requirements.txt
requests
gradio
python-dotenv
prompt_toolkit~=3.0.52